backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.32.5
openai==2.7.1        # Groq-compatible OpenAI-like client
SQLAlchemy==2.0.44
gTTS==2.5.4          # optional, only if you use TTS server-side
